spark/requirements.txt
pandas
pyarrow
numpy
notebook
google-cloud-bigquery
google-cloud-storage
wget